+++ src/restSpa/db/setup.nim
@@ -1,6 +1,16 @@
-from restSpa/db import DbConn, createTables
-from restSpa/db/models/user import newUser
+import restSpa/db
+from restSpa/db/models/user import newUser, UserRank, User
 
 proc setup*(conn: DbConn) =
   ## Creates all tables
   conn.createTables(newUser())
+
+  block createAdmin:
+    var user = newUser(
+      username = "admin",
+      email = "admin@localhost",
+      password = "admin",
+      registerIp = "",
+      rank = urAdmin
+    )
+    conn.insert user