+++ src/restSpa/routes/api/admin.nim
@@ -6,13 +6,13 @@ import pkg/prologue
 import restSpa/db
 import restSpa/db/models/user
 
-import restSpa/utils
+import restSpa/routes/utils
 
 proc r_setRank*(ctx: Context) {.async.} =
   ## Set user rank using POST
   ctx.setContentJsonHeader
   ctx.forceHttpMethod HttpPost
-  ctx.ifRank urAdmin:
+  ctx.minRank urAdmin:
     ctx.withParams(mergeGet = false):
       node.ifContains(["username", "rank"], ifContainsDefaultErr):
         let
@@ -41,27 +41,14 @@ proc r_setRank*(ctx: Context) {.async.} =
 
 proc r_getUser*(ctx: Context) {.async.} =
   ## Get all user data using POST
-  ctx.forceHttpMethod HttpPost
   ctx.setContentJsonHeader
-  ctx.withParams(mergeGet = false):
-    node.ifContains(["username", "rank"], ifContainsDefaultErr):
-      let
-        username = node{"username"}.getStr
-        rank = parseEnum[UserRank](node{"rank"}.getStr)
-      var user = newUser()
-      try:
-        inDb: dbConn.select(
-          user,
-          "User.username = ? or User.email = ?",
-          username
-        )
-      except: discard
-      if user.username.len == 0:
-        respErr "Invalid username or email"
-      else:
-        user.rank = rank
-        try:
-          inDb: dbConn.update user
-          respSuc fmt"Successfully changed user {username} rank to "
-        except DBError:
-          respErr fmt"Cannot save user: {getCurrentExceptionMsg()}"
+  ctx.forceHttpMethod HttpPost
+  ctx.minRank urAdmin:
+    ctx.withParams(mergeGet = false):
+      node.ifContains(["username"], ifContainsDefaultErr):
+        let username = node{"username"}.getStr
+        var user = User.get username
+        if user.username.len == 0:
+          respErr "Invalid username or email"
+        else:
+          respSucJson user.toJson