+++ src/restSpa/routes/api/auth.nim
@@ -3,7 +3,7 @@ import pkg/prologue
 import restSpa/db
 import restSpa/db/models/user
 
-import restSpa/utils
+import restSpa/routes/utils
 
 proc r_signIn*(ctx: Context) {.async.} =
   ## Create new user with POST
@@ -14,20 +14,14 @@ proc r_signIn*(ctx: Context) {.async.} =
       let
         username = node{"username"}.getStr
         password = node{"password"}.getStr
-      var user = newUser()
-      try:
-        inDb: dbConn.select(
-          user,
-          "User.username = ? or User.email = ?",
-          username,
-          username
-        )
-      except: discard
+      var user = User.get username
       if user.username.len == 0:
         respErr "Invalid username or email"
       elif user.password != password:
         respErr "Invalid password"
       else:
+        user.update(loginIp = ctx.request.ip)
+        inDb: dbConn.update user
         ctx.session["username"] = username
         respSuc "Success"
 
@@ -43,10 +37,17 @@ proc r_signUp*(ctx: Context) {.async.} =
         email = node{"email"}.getStr
         password = node{"password"}.getStr
       try:
-        var user = newUser(username, email, password)
+        let ip = ctx.request.ip
+        var user = newUser(
+          username = username,
+          email = email,
+          password = password,
+          registerIp = ip,
+          rank = urGhost
+        )
         inDb: dbConn.insert user
         # ctx.session["username"] = username
-        resp($(%*user))
+        respSucJson user.toJson
       except DbError:
         case getCurrentExceptionMsg():
         of "UNIQUE constraint failed: User.username":